AI_Model/requirements.txt
torch
torchvision
open_clip_torch
scikit-learn
pandas
numpy
Pillow
requests
beautifulsoup4
streamlit
timm
